Move and rename DNS and email address validators
Pure refactor that creates Validators for DNS, parsed email addresses and recipient validators in preparation of a full recipient validator.
