+++ tests/handlers/test_reset.py
@@ -1,7 +1,8 @@
 import unittest
-from rsmtpd.core.validation import EmailAddressParseResult, EmailAddressVerificationResult
 from rsmtpd.handlers.reset import ResetHandler
 from rsmtpd.handlers.shared_state import SharedState
+from rsmtpd.validators.email_address.parser import ParsedEmailAddress
+from rsmtpd.validators.email_address.recipient import ValidatedRecipient
 from tests.mocks import StubLoggerFactory, MockConfigLoader
 
 
@@ -13,8 +14,8 @@ class TestResetHandler(unittest.TestCase):
     def test_handle(self):
         handler = ResetHandler(self._mock_logger, self._mock_config_loader)
         shared_state = SharedState(("127.0.0.1", 12345))
-        shared_state.mail_from = EmailAddressParseResult()
-        shared_state.recipients = [EmailAddressVerificationResult(EmailAddressParseResult(), True)]
+        shared_state.mail_from = ParsedEmailAddress()
+        shared_state.recipients = [ValidatedRecipient(ParsedEmailAddress(), True)]
         shared_state.data_filename = "/var/tmp/rsmtpd-unit-test-test-filename"
 
         response = handler.handle("RESET", "", shared_state)
